# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - BENNU",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://bennu.test:8400"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Returns barear token",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Close the session",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Registro de usuario",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/subscription": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Crear Suscripcion",
                "description": "",
                "operationId": "eced5afd7dbdf31c2728513d41e02fc4",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type_service_id": {
                                        "description": "Id de Tipo de Servicio",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/unsubscribe/{id}": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Desuscripcion",
                "description": "",
                "operationId": "1039ff71a4500f731d1b658f6a13e059",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Estos endpoints manejan la autenticacion de los usuarios, su registro y cierre de sesion."
        },
        {
            "name": "Subscriptions",
            "description": "Estos endpoints manejan la suscripcion y desuscripcion de un usuario a n tipos de servicios."
        }
    ]
}